simplify sbg project structure
