--- project/Generator/.gitignore
@@ -1,14 +1,0 @@
-*.iml
-.gradle
-/local.properties
-/.idea/workspace.xml
-/.idea/libraries
-.DS_Store
-/build
-/captures
-
-/out
-android.jar
-bindings.txt
-my.jar
-interface-names.txt
