--- project/Generator/.idea/copyright/profiles_settings.xml
@@ -1,3 +1,0 @@
-<component name="CopyrightManager">
-  <settings default="" />
-</component>
