--- project/Generator/settings.gradle
@@ -1,1 +1,0 @@
-include ':staticbindinggenerator'
