--- project/Generator/staticbindinggenerator/.gitignore
@@ -1,1 +1,0 @@
-/build
