--- project/Generator/staticbindinggenerator/build.gradle
@@ -1,25 +1,0 @@
-apply plugin: 'java'
-
-dependencies {
-    compile 'org.apache.bcel:bcel:5.2'
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.+'
-    testCompile 'commons-io:commons-io:2.5'
-}
-
-jar {
-    manifest {
-        attributes("Manifest-Version": "1.0",
-                "Main-Class": "org.nativescript.staticbindinggenerator.Main")
-    }
-
-    from {
-        configurations.runtime.collect {
-            it.isDirectory() ? it : zipTree(it)
-        }
-
-        configurations.compile.collect {
-            it.isDirectory() ? it : zipTree(it)
-        }
-    }
-}
