--- project/Generator/staticbindinggenerator/src/test/java/com/tns/JavaScriptImplementation.java
@@ -1,9 +1,0 @@
-package com.tns;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface JavaScriptImplementation {
-    String javaScriptFile();
-}
