--- project/Generator/staticbindinggenerator/src/test/java/com/tns/NativeScriptHashCodeProvider.java
@@ -1,6 +1,0 @@
-package com.tns;
-
-public interface NativeScriptHashCodeProvider {
-    int hashCode__super();
-    boolean equals__super(Object other);
-}
