--- project/Generator/staticbindinggenerator/src/test/java/com/tns/Runtime.java
@@ -1,8 +1,0 @@
-package com.tns;
-
-public class Runtime {
-    public static void initInstance(Object instance) {}
-    public static Object callJSMethod(Object javaObject, String methodName, Class<?> retType, Object... args) {
-        return null;
-    }
-}
