--- project/Generator/staticbindinggenerator/src/test/java/org/mdkt/compiler/Main.java
@@ -1,10 +1,0 @@
-package java.org.mdkt.compiler;
-
-/**
- * Created by slavchev on 6/17/2016.
- */
-public class Main {
-    public static void main(String[] args) {
-
-    }
-}
