+++ project/staticbindinggenerator/.gitignore
@@ -1,0 +1,8 @@
+/build
+*.iml
+.gradle
+/local.properties
+/.idea/workspace.xml
+/.idea/libraries
+/a
+!.gitignore
