+++ project/staticbindinggenerator/.idea/misc.xml
