+++ project/staticbindinggenerator/src/main/java/org/nativescript/staticbindinggenerator/DataRow.java
