+++ project/staticbindinggenerator/src/main/java/org/nativescript/staticbindinggenerator/Generator.java
