+++ project/staticbindinggenerator/src/main/java/org/nativescript/staticbindinggenerator/Writer.java
