+++ project/staticbindinggenerator/src/test/java/com/tns/NativeScriptHashCodeProvider.java
@@ -1,0 +1,6 @@
+package com.tns;
+
+public interface NativeScriptHashCodeProvider {
+    int hashCode__super();
+    boolean equals__super(Object other);
+}
