+++ project/staticbindinggenerator/src/test/java/org/mdkt/compiler/CompiledCode.java
