+++ project/staticbindinggenerator/src/test/java/org/mdkt/compiler/DynamicClassLoader.java
