+++ project/staticbindinggenerator/src/test/java/org/mdkt/compiler/InMemoryJavaCompiler.java
