+++ project/staticbindinggenerator/src/test/java/org/mdkt/compiler/SourceCode.java
