+++ project/staticbindinggenerator/src/test/resources/org/nativescript/staticbindinggenerator/test/datarow-named-extend.txt
